requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
PyMuPDF
pandas
openpyxl
xlrd
lxml
pytesseract
Pillow
chromadb-client
numpy
